requirements.txt
Django==2.1.5
django-cors-headers==2.2.0
djangorestframework==3.8.2
django-rest-swagger==2.2.0
